+++ torchao/quantization/prototype/qat/utils.py
@@ -181,85 +181,6 @@ def _choose_qparams_per_token_asymmetric(
 
     return scale.to(scales_precision), zero_point.to(zero_points_precision)
 
-def _forward_pre_hook_handler(
-    mod: torch.nn.Linear,
-    prehook: Callable,
-    handler: torch.utils.hooks.RemovableHandle,
-):
-    """
-    Store a 2-tuple (prehook function, handler) as an attribute on the given linear module.
-    """
-    setattr(mod, _QAT_LINEAR_SUBCLASS_INPUT_PREHOOK, (prehook, handler))
-
-def _unwrap_affine_fake_quantized_tensor(t: torch.Tensor):
-    """
-    Return the original, non-fake-quantized float tensor from a `AffineFakeQuantizedTensor`.
-    """
-    # avoid circular dependencies
-    from torchao.quantization.prototype.qat.affine_fake_quantized_tensor import (
-        AffineFakeQuantizedTensor,
-    )
-    assert isinstance(t, AffineFakeQuantizedTensor)
-    return t.original_tensor
-
-def _is_linear_with_fq_weight(mod: torch.nn.Module, *args):
-    """
-    Return whether this is a nn.Linear module with `AffineFakeQuantizeTensor` weights.
-    """
-    # avoid circular dependencies
-    from torchao.quantization.prototype.qat.affine_fake_quantized_tensor import (
-        AffineFakeQuantizedTensor,
-    )
-    if not isinstance(mod, torch.nn.Linear) or not hasattr(mod, "weight"):
-        return False
-    weight = mod.weight
-    return isinstance(weight, AffineFakeQuantizedTensor)
-
-def _enable_fake_quant(mod: torch.nn.Module, enable: bool):
-    """
-    Enable or disable fake quantization in the activations and weights of a `nn.Linear` module.
-    """
-    from torchao.quantization.prototype.qat.affine_fake_quantized_tensor import (
-        AffineFakeQuantizedTensor,
-    )
-    if not _is_linear_with_fq_weight(mod):
-        return
-    weight = mod.weight
-    assert isinstance(weight, AffineFakeQuantizedTensor)
-    weight.fake_quant_enabled = enable
-
-    # Enable/disable input fake quant
-    if hasattr(mod, _QAT_LINEAR_SUBCLASS_INPUT_PREHOOK):
-        (prehook, handle) = getattr(mod, _QAT_LINEAR_SUBCLASS_INPUT_PREHOOK)
-        if enable and handle is None:
-            handle = mod.register_forward_pre_hook(prehook)
-        elif not enable and handle is not None:
-            handle.remove()
-            handle = None
-        setattr(mod, _QAT_LINEAR_SUBCLASS_INPUT_PREHOOK, (prehook, handle))
-
-def _get_qat_linear_subclass_inserter(
-    weight_constructor: Callable,
-    input_constructor: Optional[Callable] = None,
-) -> Callable:
-    """
-    Return a function that inserts wraps the weight and/or input activation of a
-    linear module in tensor subclasses.
-
-    Args:
-        weight_constructor: constructor of the weight subclass, accepts a tensor
-        input_constructor: (optional) constructor of the input subclass, accepts a tensor
-    """
-    def insert_subclass(lin):
-        lin.weight = torch.nn.Parameter(weight_constructor(lin.weight), requires_grad=True)
-        if input_constructor is not None:
-            prehook = lambda _, args: tuple([input_constructor(args[0])] + list(args[1:]))
-            handle = lin.register_forward_pre_hook(prehook)
-            setattr(lin, _QAT_LINEAR_SUBCLASS_INPUT_PREHOOK, (prehook, handle))
-        return lin
-
-    return insert_subclass
-
 def _get_qmin_qmax(n_bit: int):
     qmin = -(2 ** (n_bit - 1))
     qmax = 2 ** (n_bit - 1) - 1
